Move EnergyEvseServer and EnergyEvseModeServer
