+++ src/matterbridgeBehaviors.ts
@@ -69,8 +69,8 @@ import { RvcOperationalStateServer } from '@matter/main/behaviors/rvc-operationa
 import { ServiceAreaServer } from '@matter/main/behaviors/service-area';
 import { WaterHeaterModeServer } from '@matter/main/behaviors/water-heater-mode';
 import { WaterHeaterManagementServer } from '@matter/main/behaviors/water-heater-management';
-import { EnergyEvseServer } from '@matter/main/behaviors/energy-evse';
-import { EnergyEvseModeServer } from '@matter/main/behaviors/energy-evse-mode';
+// import { EnergyEvseServer } from '@matter/main/behaviors/energy-evse';
+// import { EnergyEvseModeServer } from '@matter/main/behaviors/energy-evse-mode';
 
 // AnsiLogger module
 import { AnsiLogger } from './logger/export.js';
@@ -753,36 +753,6 @@ export class MatterbridgeWaterHeaterModeServer extends WaterHeaterModeServer {
     device.changeToMode({ newMode });
     this.state.currentMode = newMode;
     device.log.info(`MatterbridgeWaterHeaterModeServer changeToMode called with newMode ${newMode} => ${supported.label}`);
-    return { status: ModeBase.ModeChangeStatus.Success, statusText: 'Success' };
-  }
-}
-
-/** ********************************************* evse  **********************************************************/
-
-export class MatterbridgeEnergyEvseServer extends EnergyEvseServer {
-  override enableCharging(): MaybePromise {
-    const device = this.endpoint.stateOf(MatterbridgeServer).deviceCommand;
-    device.enableCharging();
-    device.log.info(`MatterbridgeEnergyEvseServer enableCharging called`);
-  }
-  override disable(): MaybePromise {
-    const device = this.endpoint.stateOf(MatterbridgeServer).deviceCommand;
-    device.disable();
-    device.log.info(`MatterbridgeEnergyEvseServer disable called`);
-  }
-}
-
-export class MatterbridgeEnergyEvseModeServer extends EnergyEvseModeServer {
-  override changeToMode({ newMode }: ModeBase.ChangeToModeRequest): MaybePromise<ModeBase.ChangeToModeResponse> {
-    const device = this.endpoint.stateOf(MatterbridgeServer).deviceCommand;
-    const supported = this.state.supportedModes.find((mode) => mode.mode === newMode);
-    if (!supported) {
-      device.log.error(`MatterbridgeEnergyEvseModeServer changeToMode called with unsupported newMode: ${newMode}`);
-      return { status: ModeBase.ModeChangeStatus.UnsupportedMode, statusText: 'Unsupported mode' };
-    }
-    device.changeToMode({ newMode });
-    this.state.currentMode = newMode;
-    device.log.info(`MatterbridgeEnergyEvseModeServer changeToMode called with newMode ${newMode} => ${supported.label}`);
     return { status: ModeBase.ModeChangeStatus.Success, statusText: 'Success' };
   }
 }
